Refactor subnav component
